more build dependecies - python especially

+++ externalLibs/install_opencv.sh
@@ -12,13 +12,24 @@ set -ex
 # default: install CUDA libraries (run 'with_cuda=false ./install_opencv.sh')
 if ${with_cuda:-true} ; then
   # If CUDA is desired:
-  sudo apt-get install libxi-dev libxmu-dev freeglut3-dev build-essential binutils-gold
-  sudo ln -s /usr/lib/x86_64-linux-gnu/libglut.so /usr/lib/
-  wget http://developer.download.nvidia.com/compute/cuda/5_0/rel-update-1/installers/cuda_5.0.35_linux_64_ubuntu11.10-1.run .
-  chmod +x cuda_5.0.35_linux_64_ubuntu11.10-1.run
-  sudo ./cuda_5.0.35_linux_64_ubuntu11.10-1.run
-  echo 'export PATH=$PATH:/usr/local/cuda-5.0/bin' >> ~/.bash_profile
-  echo 'export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/cuda-5.0/lib64:/lib' >> ~/.bash_profile 
+  sudo apt-get install --yes libxi-dev libxmu-dev freeglut3-dev build-essential binutils-gold
+  if ! readlink -e /usr/lib/x86_64-linux-gnu/libglut.so ; then
+    sudo ln -s /usr/lib/x86_64-linux-gnu/libglut.so /usr/lib/
+  fi
+  cuda=cuda_5.0.35_linux_64_ubuntu11.10-1.run
+  cuda_path=/usr/local/cuda-5.0
+  if ! [ -d ${cuda_path}/lib64 ] ; then
+    if ${cuda_from_s3:-true} ; then
+      sudo pip install awscli
+      aws s3 cp s3://neon-apt-us-east-1/${cuda} .
+    else
+      wget http://developer.download.nvidia.com/compute/cuda/5_0/rel-update-1/installers/${cuda} .
+    fi
+    chmod +x ./${cuda}
+    sudo ./${cuda} -silent -toolkit -toolkitpath=${cuda_path}
+    echo "export PATH=\$PATH:${cuda_path}/bin" >> ~/.bash_profile
+    echo "export LD_LIBRARY_PATH=\$LD_LIBRARY_PATH:${cuda_path}/lib64:/lib" >> ~/.bash_profile 
+  fi
   source ~/.bash_profile
 fi
